More code error fixes

+++ applications/plugins/calculator/tinyexpr.c
@@ -278,10 +278,11 @@ void next_token(state* s) {
             s->type = TOK_NUMBER;
         } else {
             /* Look for a variable or builtin function call. */
-            if(isalpha(s->next[0])) {
+            if(isalpha((int)s->next[0])) {
                 const char* start;
                 start = s->next;
-                while(isalpha(s->next[0]) || isdigit(s->next[0]) || (s->next[0] == '_')) s->next++;
+                while(isalpha((int)s->next[0]) || isdigit((int)s->next[0]) || (s->next[0] == '_'))
+                    s->next++;
 
                 const te_variable* var = find_lookup(s, start, s->next - start);
                 if(!var) var = find_builtin(start, s->next - start);
